+++ Changelog.md
@@ -3,6 +3,12 @@
 * add Arithmetic lib;
 
 ### 2022.02.25
-* add -,*,/ operator for Vector and VectorExpr;
-* add SubFun,MulFun,DivFun class;
-* add operator-,*,/ tests for Vector;
+* add: -,*,/ operator for Vector and VectorExpr;
+* add: SubFun,MulFun,DivFun class;
+* add: operator-,*,/ tests for Vector;
+
+### 2022.02.27
+* add: Matrix,MatrixExpr class;
+* add: AddFun,SubFun,MulFun,DivFun class for Matrix and MatrixExpr;
+* add: operator+,-,*,/ for Matrix;
+* add: tests for Matrix;